Factor out form.tsx.
